move time check to utils
